% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcompile.R
\name{qcompile}
\alias{qcompile}
\title{Question Compiler}
\usage{
qcompile(folder_name)
}
\arguments{
\item{folder_name}{Name of folder containing markdowns to compile}
}
\value{
Text output of compiled code
}
\description{
This function compiles the question, answer list,
and solution from each markdown file into if else statements.
}
